+++ caom2/caom2/observation.py
@@ -459,6 +459,9 @@ class Observation(AbstractCaomEntity):
         caom_util.type_check(value, datetime, 'meta_release')
         self._meta_release = value
 
+    def validate(self, deep=True):
+        pass
+
 
 class Algorithm(CaomObject):
     """
